Continuous Deployment with CircleCI
Builds and deploys master automatically to the live website.
Makes it therefore easier to update the website.
Staging has been consciously removed since it is easy to run the
Hugo Server locally to preview changes,
which makes staging unnecessary for the website in my opinion.
Instructions for installing and running Hugo
have been added to the documentation.